cv_ats_optimizer/requirements.txt
streamlit
python-docx
pdfplumber
PyPDF2
nltk
python-dotenv
regex
